+++ app/src/main/java/com/autilite/weightlifttracker/program/session/SetSession.java
@@ -52,7 +52,7 @@ public class SetSession implements Parcelable {
         return weight;
     }
 
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
